[Refactor] #4689 floor-town.cpp/h をtown-info.cpp/h に改名した
町定義構造体があまりにもスカスカだが、将来的に対応する
