+++ src/Makefile.am
@@ -219,7 +219,6 @@ hengband_SOURCES = \
 	floor/floor-save.cpp floor/floor-save.h \
 	floor/floor-save-util.cpp floor/floor-save-util.h \
 	floor/floor-streams.cpp floor/floor-streams.h \
-	floor/floor-town.h floor/floor-town.cpp \
 	floor/floor-util.cpp floor/floor-util.h \
 	floor/geometry.cpp floor/geometry.h \
 	floor/line-of-sight.cpp floor/line-of-sight.h \
@@ -951,6 +950,7 @@ hengband_SOURCES = \
 	\
 	system/floor/floor-info.cpp system/floor/floor-info.h \
 	system/floor/floor-list.cpp system/floor/floor-list.h \
+	system/floor/town-info.cpp system/floor/town-info.h \
 	\
 	system/monrace/monrace-allocation.cpp system/monrace/monrace-allocation.h \
 	system/monrace/monrace-definition.cpp system/monrace/monrace-definition.h \
