+++ src/birth/character-builder.cpp
@@ -17,7 +17,6 @@
 #include "birth/game-play-initializer.h"
 #include "birth/quick-start.h"
 #include "core/window-redrawer.h"
-#include "floor/floor-town.h"
 #include "floor/wild.h"
 #include "game-option/option-flags.h"
 #include "io/write-diary.h"
@@ -36,6 +35,7 @@
 #include "player/race-info-table.h"
 #include "store/store-owners.h"
 #include "store/store.h"
+#include "system/floor/town-info.h"
 #include "system/player-type-definition.h"
 #include "system/redrawing-flags-updater.h"
 #include "term/gameterm.h"
