+++ src/cmd-io/cmd-dump.cpp
@@ -13,7 +13,6 @@
 #include "cmd-io/feeling-table.h"
 #include "core/asking-player.h"
 #include "dungeon/quest.h"
-#include "floor/floor-town.h"
 #include "io-dump/dump-remover.h"
 #include "io-dump/dump-util.h"
 #include "io/files-util.h"
@@ -29,6 +28,7 @@
 #include "system/angband-system.h"
 #include "system/dungeon/dungeon-definition.h"
 #include "system/floor/floor-info.h"
+#include "system/floor/town-info.h"
 #include "system/inner-game-data.h"
 #include "system/player-type-definition.h"
 #include "term/gameterm.h"
