+++ src/core/turn-compensator.cpp
@@ -1,10 +1,10 @@
 #include "core/turn-compensator.h"
-#include "floor/floor-town.h"
 #include "player-info/race-types.h"
 #include "store/store-owners.h"
 #include "store/store-util.h"
 #include "store/store.h"
 #include "system/floor/floor-info.h"
+#include "system/floor/town-info.h"
 #include "system/inner-game-data.h"
 #include "system/item-entity.h"
 #include "system/player-type-definition.h"
