+++ src/floor/fixed-map-generator.cpp
@@ -3,7 +3,6 @@
 #include "artifact/fixed-art-types.h"
 #include "dungeon/quest.h"
 #include "floor/floor-object.h"
-#include "floor/floor-town.h"
 #include "floor/wild.h"
 #include "grid/feature.h"
 #include "grid/grid.h"
@@ -29,6 +28,7 @@
 #include "system/baseitem/baseitem-list.h"
 #include "system/dungeon/dungeon-definition.h"
 #include "system/floor/floor-info.h"
+#include "system/floor/town-info.h"
 #include "system/grid-type-definition.h"
 #include "system/item-entity.h"
 #include "system/monrace/monrace-definition.h"
