+++ src/io-dump/random-art-info-dumper.cpp
@@ -1,9 +1,9 @@
 #include "io-dump/random-art-info-dumper.h"
-#include "floor/floor-town.h"
 #include "inventory/inventory-slot-types.h"
 #include "io/files-util.h"
 #include "perception/object-perception.h"
 #include "store/store-util.h"
+#include "system/floor/town-info.h"
 #include "system/item-entity.h"
 #include "system/player-type-definition.h"
 #include "util/angband-files.h"
