+++ src/knowledge/knowledge-inventory.cpp
@@ -8,7 +8,6 @@
 #include "core/show-file.h"
 #include "flavor/flavor-describer.h"
 #include "flavor/object-flavor-types.h"
-#include "floor/floor-town.h"
 #include "inventory/inventory-slot-types.h"
 #include "io-dump/dump-util.h"
 #include "object-enchant/special-object-flags.h"
@@ -20,6 +19,7 @@
 #include "sv-definition/sv-amulet-types.h"
 #include "sv-definition/sv-protector-types.h"
 #include "sv-definition/sv-ring-types.h"
+#include "system/floor/town-info.h"
 #include "system/item-entity.h"
 #include "system/player-type-definition.h"
 #include "util/angband-files.h"
