+++ src/knowledge/knowledge-self.cpp
@@ -9,7 +9,6 @@
 #include "birth/birth-explanations-table.h"
 #include "core/show-file.h"
 #include "flavor/flavor-describer.h"
-#include "floor/floor-town.h"
 #include "info-reader/fixed-map-parser.h"
 #include "io-dump/dump-util.h"
 #include "player-info/alignment.h"
@@ -19,6 +18,7 @@
 #include "player/player-status-table.h"
 #include "player/race-info-table.h"
 #include "store/store-util.h"
+#include "system/floor/town-info.h"
 #include "system/item-entity.h"
 #include "system/player-type-definition.h"
 #include "util/angband-files.h"
