+++ src/load/quest-loader.cpp
@@ -1,7 +1,6 @@
 #include "load/quest-loader.h"
 #include "artifact/fixed-art-types.h"
 #include "dungeon/quest.h"
-#include "floor/floor-town.h"
 #include "load/angband-version-comparer.h"
 #include "load/load-util.h"
 #include "load/load-zangband.h"
@@ -10,6 +9,7 @@
 #include "system/angband-exceptions.h"
 #include "system/artifact-type-definition.h"
 #include "system/floor/floor-info.h"
+#include "system/floor/town-info.h"
 #include "system/monrace/monrace-definition.h"
 #include "system/player-type-definition.h"
 #include "util/enum-converter.h"
