+++ src/load/store-loader.cpp
@@ -1,6 +1,5 @@
 #include "load/store-loader.h"
 #include "avatar/avatar.h"
-#include "floor/floor-town.h"
 #include "load/angband-version-comparer.h"
 #include "load/item/item-loader-factory.h"
 #include "load/load-util.h"
@@ -9,6 +8,7 @@
 #include "object/object-value.h"
 #include "store/store-owners.h"
 #include "store/store.h"
+#include "system/floor/town-info.h"
 #include "system/item-entity.h"
 #include "system/player-type-definition.h"
 #include "util/object-sort.h"
