+++ src/main-win/main-win-music.cpp
@@ -5,7 +5,6 @@
 
 #include "main-win/main-win-music.h"
 #include "dungeon/quest.h"
-#include "floor/floor-town.h"
 #include "main-win/main-win-cfg-reader.h"
 #include "main-win/main-win-define.h"
 #include "main-win/main-win-mci.h"
@@ -17,6 +16,7 @@
 #include "main/sound-of-music.h"
 #include "system/dungeon/dungeon-definition.h"
 #include "system/dungeon/dungeon-list.h"
+#include "system/floor/town-info.h"
 #include "system/monrace/monrace-list.h"
 #include "term/z-term.h"
 #include "util/angband-files.h"
