+++ src/player/process-death.cpp
@@ -10,7 +10,6 @@
 #include "core/asking-player.h"
 #include "core/stuff-handler.h"
 #include "flavor/flavor-describer.h"
-#include "floor/floor-town.h"
 #include "game-option/game-play-options.h"
 #include "inventory/inventory-slot-types.h"
 #include "io/files-util.h"
@@ -22,6 +21,7 @@
 #include "store/store-util.h"
 #include "store/store.h"
 #include "system/floor/floor-info.h"
+#include "system/floor/town-info.h"
 #include "system/item-entity.h"
 #include "system/player-type-definition.h"
 #include "system/redrawing-flags-updater.h"
