+++ src/room/rooms-city.cpp
@@ -1,6 +1,5 @@
 #include "room/rooms-city.h"
 #include "floor/floor-generator.h"
-#include "floor/floor-town.h"
 #include "game-option/cheat-types.h"
 #include "grid/feature.h"
 #include "grid/grid.h"
@@ -9,6 +8,7 @@
 #include "store/store.h"
 #include "system/angband-exceptions.h"
 #include "system/floor/floor-info.h"
+#include "system/floor/town-info.h"
 #include "system/grid-type-definition.h"
 #include "system/player-type-definition.h"
 #include "system/terrain-type-definition.h"
