+++ src/save/save.cpp
@@ -14,7 +14,6 @@
 #include "save/save.h"
 #include "core/object-compressor.h"
 #include "dungeon/quest.h"
-#include "floor/floor-town.h"
 #include "floor/wild.h"
 #include "inventory/inventory-slot-types.h"
 #include "io/files-util.h"
@@ -36,6 +35,7 @@
 #include "system/baseitem/baseitem-definition.h"
 #include "system/baseitem/baseitem-list.h"
 #include "system/floor/floor-info.h"
+#include "system/floor/town-info.h"
 #include "system/item-entity.h"
 #include "system/monrace/monrace-list.h"
 #include "system/player-type-definition.h"
