+++ src/spell-kind/spells-world.cpp
@@ -9,7 +9,6 @@
 #include "core/asking-player.h"
 #include "dungeon/quest-completion-checker.h"
 #include "floor/floor-mode-changer.h"
-#include "floor/floor-town.h"
 #include "floor/geometry.h"
 #include "floor/wild.h"
 #include "game-option/birth-options.h"
@@ -28,6 +27,7 @@
 #include "system/dungeon/dungeon-list.h"
 #include "system/dungeon/dungeon-record.h"
 #include "system/floor/floor-info.h"
+#include "system/floor/town-info.h"
 #include "system/grid-type-definition.h"
 #include "system/monrace/monrace-definition.h"
 #include "system/monster-entity.h"
