+++ src/store/cmd-store.cpp
@@ -5,7 +5,6 @@
 #include "flavor/flavor-describer.h"
 #include "floor/cave.h"
 #include "floor/floor-events.h"
-#include "floor/floor-town.h"
 #include "game-option/birth-options.h"
 #include "game-option/input-options.h"
 #include "inventory/inventory-object.h"
@@ -22,6 +21,7 @@
 #include "store/store.h"
 #include "system/dungeon/dungeon-definition.h"
 #include "system/floor/floor-info.h"
+#include "system/floor/town-info.h"
 #include "system/grid-type-definition.h"
 #include "system/item-entity.h"
 #include "system/player-type-definition.h"
