+++ src/store/home.cpp
@@ -1,12 +1,12 @@
 #include "store/home.h"
 #include "avatar/avatar.h"
-#include "floor/floor-town.h"
 #include "game-option/birth-options.h"
 #include "game-option/game-play-options.h"
 #include "object/object-stack.h"
 #include "object/object-value.h"
 #include "object/tval-types.h"
 #include "store/store-util.h"
+#include "system/floor/town-info.h"
 #include "system/item-entity.h"
 #include "system/player-type-definition.h"
 #include "util/object-sort.h"
