+++ src/store/rumor.cpp
@@ -1,7 +1,6 @@
 #include "store/rumor.h"
 #include "flavor/flavor-describer.h"
 #include "flavor/object-flavor-types.h"
-#include "floor/floor-town.h"
 #include "io/files-util.h"
 #include "io/tokenizer.h"
 #include "object-enchant/special-object-flags.h"
@@ -12,6 +11,7 @@
 #include "system/dungeon/dungeon-definition.h"
 #include "system/dungeon/dungeon-list.h"
 #include "system/dungeon/dungeon-record.h"
+#include "system/floor/town-info.h"
 #include "system/item-entity.h"
 #include "system/monrace/monrace-definition.h"
 #include "system/monrace/monrace-list.h"
