+++ src/store/store.cpp
@@ -7,7 +7,6 @@
 #include "store/store.h"
 #include "core/asking-player.h"
 #include "flavor/flavor-describer.h"
-#include "floor/floor-town.h"
 #include "game-option/birth-options.h"
 #include "game-option/game-play-options.h"
 #include "io/command-repeater.h"
@@ -27,6 +26,7 @@
 #include "sv-definition/sv-lite-types.h"
 #include "sv-definition/sv-scroll-types.h"
 #include "system/floor/floor-info.h"
+#include "system/floor/town-info.h"
 #include "system/item-entity.h"
 #include "system/player-type-definition.h"
 #include "term/screen-processor.h"
