+++ src/system/floor/town-info.cpp
@@ -1,3 +1,3 @@
-#include "floor/floor-town.h"
+#include "system/floor/town-info.h"
 
 std::vector<town_type> towns_info;
